feat: 逻辑完善

+++ src/main/java/priv/zxw/dictranslate/translater/AbstractEnumDictionaryTranslater.java
@@ -5,7 +5,7 @@ import priv.zxw.dictranslate.entity.DictionaryEntity;
 public abstract class AbstractEnumDictionaryTranslater implements DictionaryTranslater {
 
     @Override
-    public DictionaryEntity translate(Integer type, Integer id) {
+    public DictionaryEntity translate(Integer type, Long id) {
         return null;
     }
 }
